re arranged the project
