+++ src/components/card/index.ts
@@ -1,0 +1,3 @@
+export { default as Card } from "./Card";
+export { default as OurEnviromentCard } from "./OurEnviromentCard";
+export { default as PriceCard } from "./PriceCard";
