+++ src/components/datas/card.ts
@@ -1,0 +1,32 @@
+export const cardData = [
+  {
+    header: "Find Keywords",
+    body: "You can search for bible keywords for a category our App",
+    icon: "http://www.evethemes.com/demo/lp/startup/demo/images/64x64x1.png",
+  },
+  {
+    header: "Bible",
+    body: "We have an Bible incorporated into the app",
+    icon: "http://www.evethemes.com/demo/lp/startup/demo/images/64x64x2.png",
+  },
+  {
+    header: "Related Verses",
+    body: "Each keywords are matched with their related verses for navigation",
+    icon: "http://www.evethemes.com/demo/lp/startup/demo/images/64x64x3.png",
+  },
+  {
+    header: "User Management",
+    body: "Effective user management. We have an effective user management system",
+    icon: "http://www.evethemes.com/demo/lp/startup/demo/images/64x64x4.png",
+  },
+  {
+    header: "Effective Payment System",
+    body: "Paystack payment system integrated into the website for effective payments",
+    icon: "http://www.evethemes.com/demo/lp/startup/demo/images/64x64x5.png",
+  },
+  {
+    header: "Favorites",
+    body: "You can bookmark favorite keywords for easier use",
+    icon: "http://www.evethemes.com/demo/lp/startup/demo/images/64x64x6.png",
+  },
+];
