+++ src/components/home/GetOffer.tsx
