+++ src/components/home/index.ts
@@ -1,1 +1,7 @@
+export { default as GetInTouch } from "./GetInTouch";
+export { default as GetOffer } from "./GetOffer";
 export { default as Hero } from "./Hero";
+export { default as OurEnviroment } from "./OurEnviroment";
+export { default as OurPriceTable } from "./OurPriceTable";
+export { default as OurServices } from "./OurServices";
+export { default as WhatWeDo } from "./WhatWeDo";
