+++ src/images/logo/index.ts
@@ -1,4 +1,3 @@
-export { default as CompassLogo } from './compass_logo.png';
-export { default as CompassReverse } from './compass_reverse.png';
-export { default as CompassTop } from './compass_top.png';
-export { default as Logo } from './alogo-2.webp';
+export { default as CompassLogo } from "./compass_logo.png";
+export { default as CompassReverse } from "./compass_reverse.png";
+export { default as CompassTop } from "./compass_top.png";
